Access control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,11 @@ use App\Post;
 
 class PostsController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -74,7 +79,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
-        // dd($post);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -109,6 +118,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
         $post->delete();
         return redirect('/posts')->with('error', 'Post Removed');
     }
